+++ ee/api/routers/core_dynamic.py
@@ -2,6 +2,7 @@ from typing import Optional, Union
 
 from decouple import config
 from fastapi import Body, Depends, BackgroundTasks, Request
+from fastapi import HTTPException, status
 from starlette.responses import RedirectResponse, FileResponse
 
 import schemas
@@ -13,6 +14,7 @@ from chalicelib.core import tenants, users, projects, license
 from chalicelib.core import webhook
 from chalicelib.core.collaboration_slack import Slack
 from chalicelib.utils import SAML2_helper
+from chalicelib.utils import captcha
 from chalicelib.utils import helper
 from chalicelib.utils.TimeUTC import TimeUTC
 from or_dependencies import OR_context, OR_scope
@@ -27,9 +29,7 @@ public_app, app, app_apikey = get_routers()
 async def get_all_signup():
     return {"data": {"tenants": tenants.tenants_exists(),
                      "sso": SAML2_helper.is_saml2_available(),
-                     # "ssoProvider": SAML2_helper.get_saml2_provider(),
-                     # TODO: enable after xmlsec fix
-                     "ssoProvider": None,
+                     "ssoProvider": SAML2_helper.get_saml2_provider(),
                      "enforceSSO": config("enforce_SSO", cast=bool, default=False) and helper.is_saml2_available(),
                      "edition": license.EDITION}}
 
@@ -39,6 +39,42 @@ if config("MULTI_TENANTS", cast=bool, default=False) or not tenants.tenants_exis
     @public_app.put('/signup', tags=['signup'])
     async def signup_handler(data: schemas.UserSignupSchema = Body(...)):
         return signup.create_tenant(data)
+
+
+@public_app.post('/login', tags=["authentication"])
+async def login_user(data: schemas.UserLoginSchema = Body(...)):
+    if helper.allow_captcha() and not captcha.is_valid(data.g_recaptcha_response):
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Invalid captcha."
+        )
+
+    r = users.authenticate(data.email, data.password)
+    if r is None:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="You’ve entered invalid Email or Password."
+        )
+    if "errors" in r:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail=r["errors"][0]
+        )
+
+    r["smtp"] = helper.has_smtp()
+    content = {
+        'jwt': r.pop('jwt'),
+        'data': {
+            "user": r
+        }
+    }
+
+    return content
+
+
+@app.get('/logout', tags=["login", "logout"])
+async def logout_user(context: schemas.CurrentContext = Depends(OR_context)):
+    return {"data": "success"}
 
 
 @app.get('/account', tags=['accounts'])
@@ -162,7 +198,8 @@ async def search_sessions_by_metadata(key: str, value: str, projectId: Optional[
 
 @app.get('/projects', tags=['projects'])
 async def get_projects(context: schemas.CurrentContext = Depends(OR_context)):
-    return {"data": projects.get_projects(tenant_id=context.tenant_id, gdpr=True, user_id=context.user_id)}
+    return {"data": projects.get_projects(tenant_id=context.tenant_id, gdpr=True,
+                                          recorded=True, user_id=context.user_id)}
 
 
 # for backward compatibility