+++ tracker/tracker-mobx/src/index.ts
@@ -1,25 +1,24 @@
-import { spy } from 'mobx';
 import { App, Messages } from '@openreplay/tracker';
 
-import log from './log';
-import { Encoder, sha1 } from './syncod';
+import log from './log.js';
+import { Encoder, sha1 } from './syncod/index.js';
 
 export interface Options {
-  predicate: (ev: any) => boolean,
-  action: boolean;
-  reaction: boolean;
-  transaction: boolean;
-  compute: boolean;
+  predicate: (ev: { type: string; name: string; object: any; debugObjectName: string }) => boolean;
+  sanitize: (ev: { state: any; type: string; property: string }) => { state: any; type: string; property: string };
+  update: boolean;
+  delete: boolean;
+  add: boolean;
 }
 
-export default function(opts: Partial<Options> = {}) {
+export default function (opts: Partial<Options> = {}) {
   const options: Options = Object.assign(
     {
       predicate: () => true,
-      action: true,
-      reaction: true,
-      transaction: true,
-      compute: true,
+      sanitize: (ev) => ev,
+      update: true,
+      delete: true,
+      add: true,
     },
     opts,
   );
@@ -28,20 +27,23 @@ export default function(opts: Partial<Options> = {}) {
       return;
     }
     const encoder = new Encoder(sha1, 50);
-    spy(app.safe(ev => {
-        if (!options.predicate(ev)) return;
-        const { type } = ev;
-        ev = options[type] && log[type] && log[type](ev);
-        if (!ev) return;
-        if (typeof ev.name === 'string' && ev.name) {
-          app.send(Messages.StateAction(ev.name));
-        }
 
-        const payload = encoder.encode(ev);
-        const table = encoder.commit();
-        for (let key in table) app.send(Messages.OTable(key, table[key]));
-        app.send(Messages.MobX(type, payload));
-      }),
-    );
+    return (ev: { type: string; name: string; object: any; debugObjectName: string }) => {
+      if (!options.predicate(ev)) return;
+      const { type } = ev;
+      const event = options[type] && log[type] && log[type](ev);
+      if (!event) return;
+      const sanitizedEvent = options.sanitize(event);
+      const evType = ev.debugObjectName?.split('@')[0] || ev.type;
+      if (evType) {
+        app.send(Messages.StateAction(evType));
+      }
+
+      const payload = encoder.encode(sanitizedEvent);
+      const table = encoder.commit();
+
+      for (let key in table) app.send(Messages.OTable(key, table[key]));
+      app.send(Messages.MobX(evType, payload));
+    }
   };
 }