+++ tracker/tracker/CHANGELOG.md
@@ -1,3 +1,11 @@
+# 7.0.3
+
+- Prevent auto restart after manual stop
+
+# 7.0.2
+
+- fixed header sanitization for axios causing empty string in some cases
+
 # 7.0.1
 
 - fix time inputs capturing